etl refactoring
