+++ app/database_etl/airtable_records_handler/__init__.py
@@ -1,0 +1,3 @@
+from .airtable_records_getter import get_all_records
+from .airtable_records_formatter import apply_min_risk_of_bias, apply_study_max_estimate_grade,\
+    standardize_airtable_data
