+++ app/database_etl/postgres_tables_handler/postgres_utils.py
@@ -1,6 +1,10 @@
+import logging
+
+from sqlalchemy import distinct, func
+
 from app.serotracker_sqlalchemy import db_session, DashboardSource, \
     Country
-from sqlalchemy import distinct, func
+from app.utils.notifications_sender import send_slack_message
 
 
 def get_filter_static_options():
@@ -80,3 +84,24 @@ def get_all_filter_options():
         options["updated_at"] = session.query(func.max(DashboardSource.publication_date))[0][0].isoformat()
 
         return options
+
+
+# Send alert email if filter options have changed
+def check_filter_options(dashboard_source):
+    curr_filter_options = get_filter_static_options()
+    to_ignore = set(["All", "Multiple groups", "Multiple populations", "Multiple Types", None])
+    changed_filter_options = {}
+
+    for filter_type in curr_filter_options:
+        # Get new options for each filter type
+        new_options = set(dashboard_source[filter_type].unique())
+        # Remove options that are unused (e.g. "All", "Multiple groups", etc)
+        new_options = set([s for s in new_options if s not in to_ignore])
+        # Check to see if the new options are equal to the curr hardcoded options
+        # Check to see if the new options are equal to the curr hardcoded options
+        if new_options != set(curr_filter_options[filter_type]):
+            changed_filter_options[filter_type] = new_options
+            logging.info(new_options)
+    if len(changed_filter_options.keys()) > 0:
+        body = f"New filter options found in ETL: {changed_filter_options}"
+        send_slack_message(message=body, channel='#dev-logging-etl')
