+++ app/utils/__init__.py
@@ -4,5 +4,4 @@ from .cached_json_handler import write_to_json, read_from_json
 from .helper_funcs import validate_request_input_against_schema, convert_start_end_dates
 from .airtable_fields_config import airtable_fields_config, full_airtable_fields
 from .get_filtered_records import get_filtered_records, get_paginated_records
-from .get_filter_options import get_all_filter_options, get_filter_static_options
 from .estimate_prioritization import get_prioritized_estimates
